refactor: Code refactoring
relates #248
